% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecsades.R
\docType{package}
\name{ecsades-package}
\alias{ecsades}
\alias{ecsades-package}
\title{ecsades: Environmental contours for safe design of ships and other marine structures}
\description{

}
\details{
This package contains functions to fit joint probability distributions to bivariate wave data (wave height and 
wave period), estimate environmental contours based on the fitted distribution or directly on a given sample 
data, output the coordinates of the vertices of the contours, and generate a plot of the contours.

The choices for the joint probability distributions include the Heffernan-Tawn model \code{\link{fit_ht}} and
the Weibull-log-normal distribution \code{\link{fit_wln}}. The choices for the environmental contours are the
direct sampling contours \code{\link{estimate_dsc}}, IFORM
contours \code{\link{estimate_iform}}, generalised joint exceedance contours \code{\link{estimate_gje}}, and the
isodensity contours \code{\link{estimate_iso}}.

This package has been developed as part of the
\href{https://www.dnvgl.com/technology-innovation/sri/maritime-transport/ecsades-project.html}{ECSADES} project
and is free to use. The
statistical models and contour estimation methods included in this package are discussed in detail in the project
paper Ross et al. (2018).
}
\examples{
# Load sample data from NOAA's WaveWatch III project
data(noaa_ww3)

# Fit the Heffernan-Tawn model to the data
ht = fit_ht(
  data = noaa_ww3,
  npy = nrow(noaa_ww3)/10,
  p_margin_thresh = 0.95,
  p_dep_thresh = 0.95)

# Estimate the DSC contours using the Hs return levels as 
dsc = estimate_dsc(jdistr = ht,  output_rp = c(1,10,100))
  
# Plot output and save to a file
plot_ec(
  ec = dsc, raw_data = noaa_ww3, hs_x = FALSE, save_to_file = "dsc_ht.png",
  width = 6, height = 6, units = "in")

}
\references{
Emma Ross, Ole Christian Astrup, Elzbieta BitnerGregersen, Nigel Bunn, Graham Feld, Ben Gouldby, Arne Huseby,
Ye Liu, David Randell, Erik Vanem, Philip Jonathan, 2018. On environmental contours for marine and coastal design.
(submitted)
}
