% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_dsc.r
\name{estimate_dsc}
\alias{estimate_dsc}
\title{Estimating the direct sampling (DSC) environmental contours}
\usage{
estimate_dsc(object, output_rp, npy = NULL, n_point = 100)
}
\arguments{
\item{object}{either a fitted joint distribution of class \code{wln} or \code{ht}, or an existing sample data
set containing \code{hs} and \code{tp} as columns. See details.}

\item{output_rp}{the required return periods (in years) for the estimated contours.}

\item{npy}{the (optional) number of data points per year. This argument is only needed if the input
\code{object} is a sample data set.}

\item{n_point}{the number of points to output around each contour. The default value is 100.}
}
\value{
A set of estimated DSC contours with the specified return periods in the format
of a \code{data.table} with \code{rp}, \code{hs}, and \code{tp} as columns.
}
\description{
This function estimates the direct sampling (DSC) environmental contours with standardization, as proposed in Huesby et al. (2015),
for a given sample data set or a fitted joint distribution of class \code{ht} or \code{wln}.
}
\details{
The direct sampling contour is estimated in the origianl physical scales of \eqn{hs} and \eqn{tp} by direct
Monte Carlo simulations rather than applying the RosenBlatt transformation.

The function can be applied to an existing sample set, which can be generated by function \code{\link{sample_jdistr}}
or otherwise by importing a CSV file using function \code{\link[data.table]{fread}}. The input sample data
object must contain \code{hs} and \code{tp} as columns. The user must also specify the number of points per year
\code{npy} and a valid \code{output_rp}, i.e. the maximum \code{output_rp} must be shorter than the total
duration of the data set.

Alternatively this function can be applied to a fitted joint distribution object
generated by function \code{\link{fit_ht}} or \code{\link{fit_wln}}.  When applied to a
\code{wln} or \code{ht} object, the contour estimation makes use of the importance
sampling technique proposed in Huseby et. al. (2014) to improve the efficiency of the function.
}
\examples{
# Estimating DSC contours based on a fitted model
data(ww3_pk)
ht = fit_ht(data = ww3_pk, npy = nrow(ww3_pk)/10, margin_thresh_count = 50, dep_thresh_count = 500)
ec_ht = estimate_dsc(object = ht, output_rp = c(10,100,1000,10000))
plot_ec(ec = ec_ht, raw_data = ww3_pk)

# Estimating DSC contours based on a sample data set
data(ww3_ts)
ec_data = estimate_dsc(object = ww3_ts, output_rp = c(.5, 1, 2), npy = ww3_ts[, .N/10])
plot_ec(ec = ec_data, raw_data = ww3_ts)

}
\references{
Huseby, A., Vanem, E., Natvig, B., 2014. A new Monte Carlo method for environmental contour estimation.
Conference Proceedings. DOI:10.1201/b17399-286.

Huseby, A., Vanem, E., Natvig, B., 2015. Alternative environmental contours for structural reliability
analysis. Struct. Saf. 54, 32-45.
}
\seealso{
\code{\link{fit_ht}}, \code{\link{fit_wln}}, \code{\link{sample_jdistr}}, \code{\link{plot_ec}}
}
