% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_gje.r
\name{estimate_gje}
\alias{estimate_gje}
\title{Estimating the generalised joint exceedance contours}
\usage{
estimate_gje(object, output_rp, npy = NULL, n_point = 100)
}
\arguments{
\item{object}{either a fitted joint distribution as generated by function \code{\link{fit_ht}}
and \code{\link{fit_wln}}, or an existing sample wave data containing \code{hs} and \code{tp} as columns. See details.}

\item{output_rp}{the required return periods (in years) for the estimated contours.}

\item{npy}{the (optional) number of data points per year.  This argument is only needed if the input
\code{object} is a sample data set.}

\item{n_point}{the number of points to output around each contour. The default value is 100.}
}
\value{
A set of estimated environmental contours with the specified return periods in the format
of a \code{data.table} with \code{rp}, \code{hs}, and \code{tp} as columns.
}
\description{
This function estimates the generalised joint exceedance contours, as proposed in Jonathan et al. (2014), for a
given sample data or a fitted joint distribution of class \code{ht} or \code{wln} object.
}
\details{
The function can be applied to an existing sample data. To do this, the user needs
to supply both the sample data \code{sample_data} and the corresponding number of data points per year
\code{sample_data_npy}. The sample data can be generated using function \code{\link{sample_jdistr}} or otherwise, 
but must contain \code{hs} and \code{tp} as columns.  This \code{data.table} can also be generated
by reading a CSV file using function \code{\link[data.table]{fread}}.  

Alternatively this function can also be applied to a fitted joint distribution 
generated by function \code{\link{fit_ht}} or \code{\link{fit_wln}}.  When applied to a
\code{\link{fit_wln}} or \code{link{fit_ht}} object, the contour estimation makes use of the importance
sampling technique proposed in Huseby et. al. (2014) to improve the efficiency of the function.

For each output return period \eqn{T}, the corresponding contour is a collection of points
\code{(hs, tp)} such that the joint exceedance probability \eqn{P(Hs>hs & Tp>tp)} or
\eqn{P(Hs>hs & Tp<tp)} is constant and equal to \eqn{1/T} for the upper or the lower part of the
contour respectively.

The \eqn{T}-year contour passes through an an anchor point \code{(hs0, tp0)} where \code{hs0} is the
(\eqn{T/2})-year wave heights, and \code{tp0} is the conditional median of the
wave period given the wave height is \code{hs0}.  The joint
exceedance probability for the upper and lower half of the contour is respectively
\deqn{P(Hs>hs0 & Tp>tp0) = P(Tp>tp0 | Hs>hs0)*P(Hs>hs0) = 0.5*(2/T) = 1/T}
\deqn{P(Hs>hs0 & Tp<tp0) = P(Tp<tp0 | Hs>hs0)*P(Hs>hs0) = 0.5*(2/T) = 1/T}

More generalised forms of reference points are to be implemented.
}
\examples{
# Load data
data(ww3_pk)

# Generate a 2000-year sample data based on fitted HT
ht = fit_ht(data = ww3_pk, npy = nrow(ww3_pk)/10, margin_thresh_count = 100, dep_thresh_count = 100)
n_sim_year = 2000
sample_data = sample_jdistr(jdistr = ht, sim_year = n_sim_year)

# Estimate the GJE contours using the default reference points
gje_data = estimate_gje(
  sample_data = sample_data,
  sample_data_npy = nrow(sample_data)/n_sim_year,
  output_rp = c(1, 10, 100))
  
# Plot output
plot_ec(gje_data, ww3_pk)

}
\references{
Jonathan, P., Ewans, K., Flynn, J., 2014. On the estimation of ocean engineering design contours.
ASME J. Offshore Mech. Arct. Eng. 136:041101.

Huseby, A., Vanem, E., Natvig, B., 2014. A new Monte Carlo method for environmental contour estimation.
Conference Proceedings. DOI:10.1201/b17399-286.
}
\seealso{
\code{\link{fit_ht}}, \code{\link{fit_wln}}, \code{\link{sample_jdistr}}, \code{\link{plot_ec}}
}
