% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_iform.r
\name{estimate_iform}
\alias{estimate_iform}
\title{Estimating the IFORM contours}
\usage{
estimate_iform(object, npy = NULL, output_rp, n_point = 100,
  alpha0 = 0)
}
\arguments{
\item{object}{either a fitted joint distribution as generated by function \code{\link{fit_ht}}
and \code{\link{fit_wln}}, or an existing sample wave data containing \code{hs} and \code{tp} as columns. See details.}

\item{npy}{the (optional) number of data points per year.  This argument is only needed if the input
\code{object} is a sample data.}

\item{output_rp}{the required return periods (in years) for the estimated contours. Note that return periods
that are substantially larger than the data coverage may lead to a long processing time}

\item{n_point}{the number of points to output around each contour. The default value is 100.}

\item{alpha0}{the omission factor for the IFORM contours, as proposed in Winterstein et al. (1993).
The default value is 0, meaning no inflation to the estimated contours.}
}
\value{
A set of estimated environmental contours with the specified return periods in the format
of a \code{data.table} with \code{rp}, \code{hs}, and \code{tp} as columns.
}
\description{
This function estimates the Invserse FORM environmental contours for a 
given sample data or a fitted joint distribution of class \code{ht} or \code{wln} object.
}
\details{
The function can be applied to an existing sample data. To do this, the user needs
to supply both the sample data \code{sample_data} and the corresponding number of data points per year
\code{sample_data_npy}. The sample data can be generated using function \code{\link{sample_jdistr}} or otherwise, 
but must contain \code{hs} and \code{tp} as columns.  This \code{data.table} can also be generated
by reading a CSV file using function \code{\link[data.table]{fread}}.  

Alternatively this function can also be applied to a fitted joint distribution 
generated by function \code{\link{fit_ht}} or \code{\link{fit_wln}}.
}
\examples{
# Load data
data(ww3_pk)

# Fit the Weibull-log-normal distribution to hs and tp
wln = fit_wln(data = ww3_pk, npy = nrow(ww3_pk)/10)

# Estimate the GJE contours using the Hs return levels as 
iform = estimate_iform(
  object = wln,
  output_rp = c(1,10,100),
  alpha0 = 0.01)
  
# Plot output
plot_ec(iform, ww3_pk)

}
\references{
Winterstein, S. R., Ude, T. C., Cornell, C. A., Bjerager, P., Haver, S., 1993. Environmental parameters
for extreme variable: Inverse Form with omission factors. In: Proc. 6th Int. Conf. on
Structural Safety and Reliability, Innsbruck, Austria.
}
\seealso{
\code{\link{fit_ht}}, \code{\link{fit_wln}}, \code{\link{sample_jdistr}}, \code{\link{plot_ec}}
}
