% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_iso.r
\name{estimate_iso}
\alias{estimate_iso}
\title{Estimating the isodensity (ISO) environmental contours}
\usage{
estimate_iso(object, output_rp, n_point = 100)
}
\arguments{
\item{object}{the fitted joint distribution object as generated by function \code{\link{fit_ht}}
or \code{\link{fit_wln}}.  See details.}

\item{output_rp}{the required return periods (in years) for the estimated contours.}

\item{n_point}{the number of points to output around each contour. The default value is 100.}
}
\value{
A set of estimated ISO contours with the specified return periods in the format
of a \code{data.table} with \code{rp}, \code{hs}, and \code{tp} as columns.
}
\description{
This function estimates the isodensity (ISO) environmental contours for a fitted joint distribution of
class \code{ht} or \code{wln}.
}
\details{
The estimated isodensity contour for is a set of points with constant joint
probability density, equal to that at the anchor point.  The anchor point for the \eqn{T}-year contour
is the \eqn{K}-year return level on wave height \code{hs}, and the conditional median of
the wave period \code{tp}. See the guidance by DNV-GL section 3.7.2.

Due to the challenge of the non-parametric estimation of a joint denisty, the current version only
supports joint probability density estimation based on
the parametric form of the Weibull-log-normal model or the parametric part of the Heffernan-Tawn
model.  The residual distribution in the Heffernan-Tawn is approximted by a normal
distribution, rather than estimated using the kernel density.

For this same reason, this function does not support sample data as the input \code{object}, unlike
the other contour estimation functions in this package.  When applied
to a fitted \code{ht} object, the quadrant below the dependence thresholds is ignored.
}
\examples{
# Estimating ISO contours based on a fitted ht object
data(ww3_pk)
ht = fit_ht(data = ww3_pk, npy = nrow(ww3_pk)/10, margin_thresh_count = 50, dep_thresh_count = 500)
ec_ht = estimate_iso(object = ht, output_rp = c(10,100,1000,10000))
plot_ec(ec = ec_ht, raw_data = ww3_pk)

# Estimating ISO contours based on a fitted wln object
data(ww3_ts)
wln = fit_wln(data = ww3_ts, npy = nrow(ww3_ts)/10)
ec_wln = estimate_iso(object = wln, output_rp = c(10,100,1000,10000))
plot_ec(ec = ec_wln, raw_data = ww3_ts)

}
\references{
DNVGL-RP-C205, Recommended Practice - Environmental conditions and environmental loads,
Edition August 2017, https://rules.dnvgl.com/docs/pdf/DNVGL/RP/2017-08/DNVGL-RP-C205.pdf

Heffernan, Janet & A. Tawn, Jonathan. (2004). A Conditional Approach for Multivariate Extreme Values.
Journal of the Royal Statistical Society Series B. 66. 497-546. 10.1111/j.1467-9868.2004.02050.x.
}
\seealso{
\code{\link{fit_ht}}, \code{\link{fit_wln}}, \code{\link{plot_ec}}
}
