% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ht.r
\name{fit_ht}
\alias{fit_ht}
\title{Fitting Heffernan-Tawn model to wave data}
\usage{
fit_ht(data, npy, margin_thresh_count, dep_thresh_count)
}
\arguments{
\item{data}{the wave data in the form a \code{data.table} with wave height \code{hs} and wave period
\code{tp} as columns}

\item{npy}{the number of data points per year, usually estimated by the number of rows in the
supplied wave data divided by the total period of data coverage (in years)}

\item{margin_thresh_count}{an integer indicating the number of exceedances strcitly above the threshold
used for fitting the Generalised Pareto Distribution (GPD) for the upper tail of the marginal distribution
for \code{hs} and \code{tp}}

\item{dep_thresh_count}{an integer indicating the number of exceedances (of the conditioning variable)
strcitly above the threshold used for fitting the conditional multivariate extreme value model of
Heffernan-Tawn (2004)}
}
\value{
An joint distribution object of class \code{ht} containing the key information of a fitted
Heffernan-Tawn model, including the marginal distribution parameters and the dependence model parameters.
}
\description{
This function fits a Heffernan-Tawn (\code{ht}) model to the given wave data. The marginal distributions are
a mixture model with an empirical distribution below a specified threshold, and the generalised Pareto distribution
(GPD) above the threshold. The dependence structure is the conditional extreme value model proposed in
Heffernan and Tawn (2004).
}
\details{
The input \code{data} must be a \code{data.table} object with \code{hs} and \code{tp} columns. This can be
generated by reading a CSV file using function \code{\link[data.table]{fread}}.

By the asymptotic properties of the GPD and the HT model, the model bias is only negligible when the 
thresholds are sufficiently high. This function prompts the user with the exceedance probability implied by
their choie of the arguments \code{margin_thresh_count} and \code{dep_thresh_count} to remind them that
the exceedance probability needs to be quite small for the fit of GPD and HT to be valid.
}
\examples{
# Load data
data(noaa_ww3)

# Fit Heffernan-Tawn model
noaa_ht = fit_ht(data = noaa_ww3, npy = nrow(noaa_ww3)/10, margin_thresh_count = 100, dep_thresh_count = 100)

}
\references{
Heffernan, Janet & A. Tawn, Jonathan. (2004). A Conditional Approach for Multivariate Extreme Values.
Journal of the Royal Statistical Society Series B. 66. 497-546. 10.1111/j.1467-9868.2004.02050.x.
}
\seealso{
\code{\link{fit_wln}}, \code{\link{sample_jdistr}}
}
