% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa_ww3.r
\docType{data}
\name{noaa_ww3}
\alias{noaa_ww3}
\title{Sample wave data by NOAA}
\format{An object of class \code{data.table} with columns \code{time}, \code{hs} and \code{tp};
see \code{\link{data.table}}.}
\source{
\href{http://polar.ncep.noaa.gov/waves/wavewatch/}{NOAA WAVEWATCH III Model}
}
\description{
This sample dataset contains wave peak data (large Hs and/or Tp) as extracted from
the 3-hourly (before March 1st 2010) and 1-hourly (after March 1st 2010) wave time series data 
provided by NOAA's WaveWatch III project for location 30.5N, 240E.
}
\examples{
data(noaa_ww3)
}
\references{
The WAVEWATCH III Development Group., 2016:
User manual and system documentation of WAVEWATCH III version 5.16.
NOAA / NWS / NCEP / MMAB Technical Note 329, 326 pp. + Appendices.
}
