% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ec.r
\name{plot_ec}
\alias{plot_ec}
\title{Plotting the environmental contours}
\usage{
plot_ec(ec, raw_data = NULL, hs_x = TRUE, save_to_file = NULL, ...)
}
\arguments{
\item{ec}{the environmental contours in the format of a \code{data.table} with
\code{rp}, \code{hs} and \code{tp} as columns.  These can be generated using the contour
estimation functions such as \code{\link{estimate_dsc}}, \code{\link{estimate_iform}}, etc.}

\item{raw_data}{the (optional) raw data behind the the fitted contours. Providing this data
will add the corresponding points to the final plot. The default value is \code{NULL}.}

\item{hs_x}{whether or not the x-axis is the wave height \code{hs}. Setting \code{hs_x = TRUE}
means the plot is \code{(hs, tp)}; otherwise it would be \code{(tp, hs)}. The default
value is \code{TRUE}.}

\item{save_to_file}{an (optional) file name for saving the image to. The file type can be 
anything that is supported by the \code{\link{ggsave}} function. The default value is
\code{NULL} which means the image is not saved.}

\item{...}{additional (optional) arguments passed onto function \code{\link{ggsave}}.}
}
\value{
A \code{\link{ggplot}} object containing the input environmental contours
and other optional specified items.
}
\description{
This function prints on screen a given set of environmental contours and saves the image
to a file with optional customisation.
}
\examples{
# Load sample data from NOAA's WaveWatch III project
data(ww3_pk)

# Fit the Heffernan-Tawn model to the data
ht = fit_ht(
  data = ww3_pk,
  npy = nrow(ww3_pk)/10,
  margin_thresh_count = 100,
  dep_thresh_count = 500)

# Estimate the DSC contours using the Hs return levels as 
dsc = estimate_dsc(object = ht,  output_rp = c(1,10,100))
  
# Plot output and save to a file
plot_ec(
  ec = dsc, raw_data = ww3_pk, hs_x = FALSE, save_to_file = "dsc_ht.png",
  width = 8, height = 6, units = "in")

}
