% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_dsc.r
\name{estimate_dsc}
\alias{estimate_dsc}
\title{Estimating the direct sampling contours}
\usage{
estimate_dsc(object, output_rp, npy = NULL, n_point = 100)
}
\arguments{
\item{object}{either a fitted joint distribution as generated by function \code{\link{fit_ht}}
and \code{\link{fit_wln}}, or an existing sample wave data containing \code{hs} and \code{tp} as columns. See details.}

\item{output_rp}{the required return periods (in years) for the estimated contours.}

\item{npy}{the (optional) number of data points per year.  This argument is only needed if the input
\code{object} is a sample data set.}

\item{n_point}{the number of points to output around each contour. The default value is 100.}
}
\value{
A set of estimated environmental contours with the specified return periods in the format
of a \code{data.table} with \code{rp}, \code{hs}, and \code{tp} as columns.
}
\description{
This function estimates the direct sampling contours with standardization, as proposed in Huesby et al. (2015), for a given
sample data or a fitted joint distribution of class \code{ht} or \code{wln} object.
}
\details{
The function can be applied to an existing sample data. To do this, the user needs
to supply both the sample data \code{sample_data} and the corresponding number of data points per year
\code{sample_data_npy}. The sample data can be generated using function \code{\link{sample_jdistr}} or otherwise, 
but must contain \code{hs} and \code{tp} as columns.  This \code{data.table} can also be generated
by reading a CSV file using function \code{\link[data.table]{fread}}.  

Alternatively this function can also be applied to a fitted joint distribution 
generated by function \code{\link{fit_ht}} or \code{\link{fit_wln}}.  When applied to a
\code{\link{fit_wln}} or \code{link{fit_ht}} object, the contour estimation makes use of the importance sampling
technique proposed in Huseby et. al. (2014) to improve the efficiency of the function.
}
\examples{
# Load data
data(ww3_pk)

# Estimate the DSC based on fitted Heffernan-Tawn model
ht = fit_ht(data = ww3_pk, npy = nrow(ww3_pk)/10, margin_thresh_count = 100, dep_thresh_count = 100)
dsc_ht = estimate_dsc(jdistr = ht, output_rp = c(1,10,100))

# Plot output
plot_ec(dsc_ht, ww3_pk)

# Estimate the DSC based on sample data
wln = fit_wln(data = ww3_pk, npy = nrow(ww3_pk)/10)
sample_data = sample_jdistr(jdistr = wln, sim_year = 1000)
dsc_data = estimate_dsc(sample_data = sample_data, sample_data_npy = nrow(sample_data)/1000, output_rp = c(1,10,100))

# Plot output
plot_ec(dsc_data, ww3_pk)

}
\references{
Huseby, A., Vanem, E., Natvig, B., 2014. A new Monte Carlo method for environmental contour estimation.
Conference Proceedings. DOI:10.1201/b17399-286.

Huseby, A., Vanem, E., Natvig, B., 2015. Alternative environmental contours for structural reliability
analysis. Struct. Saf. 54, 32-45.
}
\seealso{
\code{\link{fit_ht}}, \code{\link{fit_wln}}, \code{\link{sample_jdistr}}, \code{\link{plot_ec}}
}
