% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_dsc.r
\name{estimate_dsc}
\alias{estimate_dsc}
\title{Estimating the direct sampling contours}
\usage{
estimate_dsc(jdistr, sample_data = NULL, sample_data_npy = NULL,
  output_rp, n_point = 100)
}
\arguments{
\item{jdistr}{the (optional) fitted joint distribution as generated by function \code{\link{fit_ht}}
or \code{\link{fit_wln}}.  See details.}

\item{sample_data}{the (optional) existing sample wave data containing \code{hs} and \code{tp} as columns. See details.}

\item{sample_data_npy}{the (optional) number of data points per year of the sample data. See details.}

\item{output_rp}{the required return periods (in years) for the estimated contours. Note that return periods
that are substantially larger than the data coverage may lead to a long processing time.}

\item{n_point}{the number of points to output around each contour. The default value is 100.}
}
\value{
A set of estimated environmental contours with the specified return periods in the format
of a \code{data.table} with \code{rp}, \code{hs}, and \code{tp} as columns.
}
\description{
This function estimates the direct sampling contours with standardization, as proposed in Huesby et al. (2015), for a given
sample data or a fitted joint distribution of class \code{ht} or \code{wln} object.
}
\details{
The input \code{sample_data} must be a \code{data.table} object with \code{hs} and \code{tp} columns. This can be
generated by reading a CSV file using function \code{\link[data.table]{fread}}.

This function can be applied to either a fitted joint distribution of class \code{ht} or \code{wln}. This
happens when the user supplies a valid \code{ht} or \code{wln} object, and ignores arguments \code{sample_data}
and \code{sample_data_npy}.

This function can also be applied to an existing sample data. To do this, the user needs
to supply both the sample data \code{sample_data} and the corresponding number of data points per year
\code{sample_data_npy}. The sample data can be generated using function \code{\link{sample_jdistr}} or otherwise, 
but must contain \code{hs} and \code{tp} as columns.

Providing \code{sample_data} and \code{sample_data_npy} will make the input \code{jdistr} redundant.
}
\examples{
# Load data
data(noaa_ww3)

# Estimate the DSC based on fitted Heffernan-Tawn model
ht = fit_ht(data = noaa_ww3, npy = nrow(noaa_ww3)/10, p_margin_thresh = 0.95, p_dep_thresh = 0.95)
dsc_ht = estimate_dsc(jdistr = ht, output_rp = c(1,10,100))

# Plot output
plot_ec(dsc_ht, noaa_ww3)

# Estimate the DSC based on sample data
wln = fit_wln(data = noaa_ww3, npy = nrow(noaa_ww3)/10)
sample_data = sample_jdistr(jdistr = wln, sim_year = 1000)
dsc_data = estimate_dsc(sample_data = sample_data, sample_data_npy = nrow(sample_data)/1000, output_rp = c(1,10,100))

# Plot output
plot_ec(dsc_data, noaa_ww3)

}
\references{
Huseby, A., Vanem, E., Natvig, B., 2015. Alternative environmental contours for structural reliability
analysis. Struct. Saf. 54, 32-45.
}
\seealso{
\code{\link{fit_ht}}, \code{\link{fit_wln}}, \code{\link{sample_jdistr}}, \code{\link{plot_ec}}
}
