% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_gje.r
\name{estimate_gje}
\alias{estimate_gje}
\title{Estimating the generalised joint exceedance contours}
\usage{
estimate_gje(jdistr, sample_data = NULL, sample_data_npy = NULL,
  output_rp, n_point = 100, ref_hs = NULL, ref_tp = NULL)
}
\arguments{
\item{jdistr}{the (optional) fitted joint distribution as generated by function \code{\link{fit_ht}}
or \code{\link{fit_wln}}.  See details.}

\item{sample_data}{the (optional) existing sample wave data containing \code{hs} and \code{tp} as columns. See details.}

\item{sample_data_npy}{the (optional) number of data points per year of the sample data. See details.}

\item{output_rp}{the required return periods (in years) for the estimated contours. Note that return periods
that are substantially larger than the data coverage may lead to a long processing time.}

\item{n_point}{the number of points to output around each contour. The default value is 100.}

\item{ref_hs, ref_tp}{the wave height and wave period for the reference point(s). When either is
missing, the default reference points are \code{ref_hs = 0} and \code{ref_tp} equal to the median of the
conditional distribution of the wave period given the wave heights are above the
(\code{output_rp/2})-year return levels.  See details.}
}
\value{
A set of estimated environmental contours with the specified return periods in the format
of a \code{data.table} with \code{rp}, \code{hs}, and \code{tp} as columns.
}
\description{
This function estimates the generalised joint exceedance contours, as proposed in Jonathan et al. (2014), for a
given sample data or a fitted joint distribution of class \code{ht} or \code{wln} object.
}
\details{
The input \code{sample_data} must be a \code{data.table} object with \code{hs} and \code{tp} columns. This
can be generated by reading a CSV file using function \code{\link[data.table]{fread}}.

For each return period of the output contours, the corresponding reference point (\code{hs}, \code{tp})
divides the space into (up to) four quadrants.  The joint exceedance region of a given point is
then the corner formed by the point as the vertex, two semi-straight lines parallel to the \code{hs}
and \code{tp} axes respectively, and facing away from the reference point. Please refer to Jonathan et al.
(2014) for the mathematical formulation of the definition.

The default reference point will divide the space into two parts.  The resulting contour will pass
through a set of anchor points \code{(hs0, tp0)} where \code{hs0} is the (\code{output_rp/2})-year
return levels for wave heights, and \code{tp0 = ref_tp}.  The joint exceedance probability calculation
is therefore 
\deqn{P(Hs>hs0 & Tp>tp0) = P(Tp>tp0 | Hs>hs0)*P(Hs>hs0) = 0.5*(2/output_rp) = 1/output_rp}
\deqn{P(Hs>hs0 & Tp<tp0) = P(Tp<tp0 | Hs>hs0)*P(Hs>hs0) = 0.5*(2/output_rp) = 1/output_rp}

This function can be applied to either a fitted joint distribution of class \code{ht} or \code{wln}. This
happens when the user supplies a valid \code{ht} or \code{wln} object, and ignores arguments \code{sample_data}
and \code{sample_data_npy}.

This function can also be applied to an existing sample data. To do this, the user needs
to supply both the sample data \code{sample_data} and the corresponding number of data points per year
\code{sample_data_npy}. The sample data can be generated using function \code{\link{sample_jdistr}} or otherwise, 
but must contain \code{hs} and \code{tp} as columns.

Providing \code{sample_data} and \code{sample_data_npy} will make the input \code{jdistr} redundant.
}
\examples{
# Load data
data(noaa_ww3)

# Generate a 2000-year sample data based on fitted HT
ht = fit_ht(data = noaa_ww3, npy = nrow(noaa_ww3)/10, margin_thresh_count = 100, dep_thresh_count = 100)
n_sim_year = 2000
sample_data = sample_jdistr(jdistr = ht, sim_year = n_sim_year)

# Estimate the GJE contours using the default reference points
gje_data = estimate_gje(
  sample_data = sample_data,
  sample_data_npy = nrow(sample_data)/n_sim_year,
  output_rp = c(1, 10, 100))
  
# Plot output
plot_ec(gje_data, noaa_ww3)

}
\references{
Jonathan, P., Ewans, K., Flynn, J., 2014. On the estimation of ocean engineering design contours.
ASME J. Offshore Mech. Arct. Eng. 136:041101.
}
\seealso{
\code{\link{fit_ht}}, \code{\link{fit_wln}}, \code{\link{sample_jdistr}}, \code{\link{plot_ec}}
}
