% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_gje.r
\name{estimate_gje}
\alias{estimate_gje}
\title{Estimating the generalised joint exceedance (GJE) environmental contours}
\usage{
estimate_gje(object, output_rp, type = 2, npy = NULL, n_point = 100)
}
\arguments{
\item{object}{either a fitted joint distribution of class \code{wln} or \code{ht}, or an existing sample data
set containing \code{hs} and \code{tp} as columns. See details.}

\item{output_rp}{the required return periods (in years) for the estimated contours.}

\item{type}{the type of joint exceedance definition. The default value is 2. See details.}

\item{npy}{the (optional) number of data points per year. This argument is only needed if the input
\code{object} is a sample data set.}

\item{n_point}{the number of points to output around each contour. The default value is 100.}
}
\value{
A set of estimated GJE contours with the specified return periods in the format
of a \code{data.table} with \code{rp}, \code{hs}, and \code{tp} as columns.
}
\description{
This function estimates the generalised joint exceedance (GJE) environmental contours, as proposed in Jonathan et al. (2014), for a
given sample data set or a fitted joint distribution of class \code{ht} or \code{wln}.
}
\details{
The joint exceedance contour is a collection of points corresponding to a constant joint exceedance proability.
The definition is one-sided or two-sided depending on the input argument \code{type}:

\itemize{

  \item If \code{type = 1}, the \eqn{T}-year contour passes through any point \eqn{(hs, tp)} such that the 
  joint exceedance probability \eqn{P(Hs > hs & Tp > tp)} is one in \eqn{T} years.
  
  \item If \code{type = 2}, the \eqn{T}-year contour passes through an anchor point \eqn{(hs0, tp0)}
  where \eqn{hs0} is the (\eqn{T/2})-year wave heights, and \eqn{tp0} is the conditional median of the
  wave period given the wave height is \eqn{hs0}. The joint exceedance probability is defined as
  \eqn{P(Hs > hs & Tp > tp)} for all \eqn{tp > tp0} and \eqn{P(Hs > hs & Tp < tp)} for all \eqn{tp < tp0}.

}

The function can be applied to an existing sample set, which can be generated by function \code{\link{sample_jdistr}}
or otherwise by importing a CSV file using function \code{\link[data.table]{fread}}. The input sample data
object must contain \code{hs} and \code{tp} as columns. The user must also specify the number of points per year
\code{npy} and a valid \code{output_rp}, i.e. the maximum \code{output_rp} must be shorter than the total
duration of the data set.

Alternatively this function can be applied to a fitted joint distribution object
generated by function \code{\link{fit_ht}} or \code{\link{fit_wln}}.  When applied to a
\code{wln} or \code{ht} object, the contour estimation makes use of the importance
sampling technique proposed in Huseby et. al. (2014) to improve the efficiency of the function.
}
\examples{
# Estimating type-1 GJE contours based on a fitted model
data(ww3_pk)
wln = fit_wln(data = ww3_pk, npy = nrow(ww3_pk)/10)
ec_wln = estimate_gje(object = wln, output_rp = c(10,100,1000,10000), type = 1)
plot_ec(ec = ec_wln, raw_data = ww3_pk)

# Estimating type-2 GJE contours based on a sample data set
data(ww3_ts)
ec_data = estimate_gje(object = ww3_ts, output_rp = c(.5, 1, 2), type = 2, npy = ww3_ts[, .N/10])
plot_ec(ec = ec_data, raw_data = ww3_ts)

}
\references{
Jonathan, P., Ewans, K., Flynn, J., 2014. On the estimation of ocean engineering design contours.
ASME J. Offshore Mech. Arct. Eng. 136:041101.

Huseby, A., Vanem, E., Natvig, B., 2014. A new Monte Carlo method for environmental contour estimation.
Conference Proceedings. DOI:10.1201/b17399-286.
}
\seealso{
\code{\link{fit_ht}}, \code{\link{fit_wln}}, \code{\link{sample_jdistr}}, \code{\link{plot_ec}}
}
