% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_gje.r
\name{estimate_gje}
\alias{estimate_gje}
\title{Estimating the generalised joint exceedance contours}
\usage{
estimate_gje(jdistr, sample_data = NULL, sample_data_npy = NULL,
  output_rp, n_point = 100, ref_tp = rep(0, length(output_rp)),
  ref_hs = rep(0, length(output_rp)))
}
\arguments{
\item{jdistr}{the (optional) fitted joint distribution as generated by function \code{\link{fit_ht}}
or \code{\link{fit_wln}}.  See details.}

\item{sample_data}{the (optional) existing sample wave data containing \code{hs} and \code{tp} as columns. See details.}

\item{sample_data_npy}{the (optional) number of data points per year of the sample data. See details.}

\item{output_rp}{the required return periods (in years) for the estimated contours. Note that return periods
that are substantially larger than the data coverage may lead to a long processing time.}

\item{n_point}{the number of points to output around each contour. The default value is 100.}

\item{ref_tp}{the wave period for the reference point. The default value is 0. See details.}

\item{ref_hs}{the wave height for the reference point. The default value is 0. See details.}
}
\value{
A set of estimated environmental contours with the specified return periods in the format
of a \code{data.table} with \code{rp}, \code{hs}, and \code{tp} as columns.
}
\description{
This function estimates the generalised joint exceedance contours, as proposed in Jonathan et al. (2014), for a
given sample data or a fitted joint distribution of class \code{ht} or \code{wln} object.
}
\details{
This function can be applied to either a fitted joint distribution of class \code{ht} or \code{wln}. This
happens when the user supplies a valid \code{ht} or \code{wln} object, and ignores arguments \code{sample_data}
and \code{sample_data_npy}.

This function can also be applied to an existing sample data. To do this, the user needs
to supply both the sample data \code{sample_data} and the corresponding number of data points per year
\code{sample_data_npy}. The sample data can be generated using function \code{\link{sample_jdistr}} or otherwise, 
but must contain \code{hs} and \code{tp} as columns.

Providing \code{sample_data} and \code{sample_data_npy} will make the input \code{jdistr} redundant.

For each return period of the output contours, the corresponding reference point (\code{hs}, \code{tp}) divides
the space into (up to) four quadrants.  The joint exceedance region of a given point
is then the corner formed by the point as the vertex, two semi-straight lines parallel to the \code{hs} and \code{tp}
axes respectively, and facing away from the reference point.

Please refer to Jonathan et al. (2014) for the mathematical formulation of the definition.
}
\examples{
# Load data
data(noaa_ww3)

# Generate a 2000-year sample data based on fitted HT
ht = fit_ht(data = noaa_ww3, npy = nrow(noaa_ww3)/10, p_margin_thresh = 0.95, p_dep_thresh = 0.95)
n_sim_year = 2000
sample_data = sample_jdistr(jdistr = ht, sim_year = n_sim_year)

# Estimate the 10, 100, 1000-year return levels of Hs based on the simulated data
out_rp = c(1, 10, 100)
rl_hs = sample_data[sort.list(hs, decreasing = T)[n_sim_year/out_rp], hs]

# Estimate the median of the Tp given Hs is equal to these return levels
ref_tp = c()
for(i in 1:3){
  ref_tp[i] = sample_data[abs(hs-rl_hs[i])<=1, median(tp)]
}

# Estimate the GJE contours using the Hs return levels as 
gje_data = estimate_gje(
  sample_data = sample_data,
  sample_data_npy = nrow(sample_data)/n_sim_year,
  output_rp = out_rp,
  ref_tp = ref_tp)
  
# Plot output
plot_ec(gje_data, noaa_ww3)

}
\references{
Jonathan, P., Ewans, K., Flynn, J., 2014. On the estimation of ocean engineering design contours.
ASME J. Offshore Mech. Arct. Eng. 136:041101.
}
\seealso{
\code{\link{fit_ht}}, \code{\link{fit_wln}}, \code{\link{sample_jdistr}}, \code{\link{plot_ec}}
}
