% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_iform.r
\name{estimate_iform}
\alias{estimate_iform}
\title{Estimating the inverse first-order reliability method (IFORM) environmental contours}
\usage{
estimate_iform(object, output_rp, npy = NULL, n_point = 100,
  alpha0 = 0)
}
\arguments{
\item{object}{either a fitted joint distribution of class \code{wln} or \code{ht}, or an existing sample data
set containing \code{hs} and \code{tp} as columns. See details.}

\item{output_rp}{the required return periods (in years) for the estimated contours.}

\item{npy}{the (optional) number of data points per year. This argument is only needed if the input
\code{object} is a sample data set.}

\item{n_point}{the number of points to output around each contour. The default value is 100.}

\item{alpha0}{the omission factor for the IFORM contours. The default value is 0, meaning no
inflation to the estimated contours.}
}
\value{
A set of estimated IFORM contours with the specified return periods in the format
of a \code{data.table} with \code{rp}, \code{hs}, and \code{tp} as columns.
}
\description{
This function estimates the Invserse FORM (IFORM) environmental contours, as proposed in Winterstein et al. (1993)
for a given sample data set or a fitted joint distribution of class \code{ht} or \code{wln}.
}
\details{
The IFORM contour is constructed based on the first-order reliability method (FORM) and 
it searches a hypersphere of constant radius to find the maximum response.

The function can be applied to an existing sample set, which can be generated by function \code{\link{sample_jdistr}}
or otherwise by importing a CSV file using function \code{\link[data.table]{fread}}. The input sample data
object must contain \code{hs} and \code{tp} as columns. The user must also specify the number of points per year
\code{npy}.

Alternatively this function can be applied to a fitted joint distribution object
generated by function \code{\link{fit_ht}} or \code{\link{fit_wln}}.  When applied to a \code{ht} object,
the estimated contour is a blending between the two conditional models \eqn{Hs|Tp>tp} and \eqn{Tp|Hs>hs} above
high thresholds \eqn{tp} and \eqn{hs}, as specified in Heffernan and Tawn (2004).
}
\examples{
# Estimating IFORM contours based on a fitted model
data(ww3_pk)
ht = fit_ht(data = ww3_pk, npy = nrow(ww3_pk)/10, margin_thresh_count = 50, dep_thresh_count = 500)
ec_ht = estimate_iform(object = ht, output_rp = c(10,100,1000,10000))
plot_ec(ec = ec_ht, raw_data = ww3_pk)

# Estimating IFORM contours based on a sample data set
data(ww3_ts)
ec_data = estimate_iform(object = ww3_ts, output_rp = c(.5, 1, 2), npy = ww3_ts[, .N/10])
plot_ec(ec = ec_data, raw_data = ww3_ts)

}
\references{
Winterstein, S. R., Ude, T. C., Cornell, C. A., Bjerager, P., Haver, S., 1993. Environmental parameters
for extreme variable: Inverse Form with omission factors. In: Proc. 6th Int. Conf. on
Structural Safety and Reliability, Innsbruck, Austria.

Heffernan, Janet & A. Tawn, Jonathan. (2004). A Conditional Approach for Multivariate Extreme Values.
Journal of the Royal Statistical Society Series B. 66. 497-546. 10.1111/j.1467-9868.2004.02050.x.
}
\seealso{
\code{\link{fit_ht}}, \code{\link{fit_wln}}, \code{\link{sample_jdistr}}, \code{\link{plot_ec}}
}
