% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_iso.r
\name{estimate_iso}
\alias{estimate_iso}
\title{Estimating the isodensity contours}
\usage{
estimate_iso(jdistr, output_rp, n_grid = 100, range_multiplier = 1.5)
}
\arguments{
\item{jdistr}{the (optional) fitted joint distribution as generated by function \code{\link{fit_ht}}
or \code{\link{fit_wln}}.  See details.}

\item{output_rp}{the required return periods (in years) for the estimated contours. Note that return periods
that are substantially larger than the data coverage may lead to a long processing time}

\item{n_grid}{the resolution used for the numerical integration of joint probability density. 
The default value is 100. See details.}

\item{range_multiplier}{the range in \code{hs} and \code{tp} used for the numerical integration
of the joint probability density, as a multiplier of the actual range of the sample data.  See details.}

\item{sample_data}{the (optional) existing sample wave data containing \code{hs} and \code{tp} as columns. See details.}

\item{sample_data_npy}{the (optional) number of data points per year of the sample data. See details.}
}
\value{
A set of estimated environmental contours with the specified return periods in the format
of a \code{data.table} with \code{rp}, \code{hs}, and \code{tp} as columns.
}
\description{
This function estimates the isodensity environmental contours for a 
given sample data or a fitted joint distribution of class \code{ht} or \code{wln} object.
}
\details{
The estimated contour for a given return period \code{rp} is a set of points with constant probability
density, such that the probability of a point lying in the interior of the contour is \code{1-1/rp}.

The probability is estimated using numerical integration over a region defined by \code{range_multiplier}
at a resolution specified by \code{n_grid}. For exmaple, setting \code{range_multiplier = 1.5} and 
\code{n_grid = 100} means the joint probability is estimated over a mesh grid of \eqn{100 * 100}
from \eqn{min(hs)/1.5} to \eqn{max(hs)*1.5} and similarly for \code{tp}. Higher \code{range_multiplier}
and \code{n_grid} mean less estimation error, but longer processign time.

The input \code{sample_data} must be a \code{data.table} object with \code{hs} and \code{tp} columns. This can be
generated by reading a CSV file using function \code{\link[data.table]{fread}}.

This function can be applied to either a fitted joint distribution of class \code{ht} or \code{wln}. This
happens when the user supplies a valid \code{ht} or \code{wln} object, and ignores arguments \code{sample_data}
and \code{sample_data_npy}.

This function can also be applied to an existing sample data. To do this, the user needs
to supply both the sample data \code{sample_data} and the corresponding number of data points per year
\code{sample_data_npy}. The sample data can be generated using function \code{\link{sample_jdistr}} or otherwise, 
but must contain \code{hs} and \code{tp} as columns.

The current implementation of the estimation method cannot efficiently deal with the lower-left quadrant
of a fitted \code{ht} object where both \code{hs} and \code{tp} are below the dependence treshold.
The resulting output contours do not include that quadrant to avoid excessive computation.

Providing \code{sample_data} and \code{sample_data_npy} will make the input \code{jdistr} redundant.
}
\examples{
# Load data
data(noaa_ww3)

# Fit the Weibull-log-normal distribution to hs and tp
wln = fit_wln(data = noaa_ww3, npy = nrow(noaa_ww3)/10)

# Estimate the GJE contours using the Hs return levels as 
iso = estimate_iso(jdistr = wln,  output_rp = c(1,10,100))
  
# Plot output
plot_ec(iform, noaa_ww3)

}
\seealso{
\code{\link{fit_ht}}, \code{\link{fit_wln}}, \code{\link{sample_jdistr}}, \code{\link{plot_ec}}
}
