% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ht.r
\name{fit_ht}
\alias{fit_ht}
\title{Fitting Heffernan-Tawn model to wave data}
\usage{
fit_ht(data, npy, p_margin_thresh, p_dep_thresh)
}
\arguments{
\item{data}{the wave data in the form a \code{data.table} with wave height \code{hs} and wave period
\code{tp} as columns}

\item{npy}{the number of data points per year, usually estimated by the number of rows in the
supplied wave data divided by the total period of data coverage (in years)}

\item{p_margin_thresh}{the threshold (in terms of the non-exceedanc probability) used for fitting the
Generalised Pareto Distribution for the upper tail of the marginal distribution for \code{hs} and \code{tp}}

\item{p_dep_thresh}{the threshold (in terms of the non-exceedance probability) used for fitting the
conditional multivariate extreme value model of Heffernan-Tawn (2004)}
}
\value{
An joint distribution object of class \code{ht} containing the key information of a fitted
Heffernan-Tawn model, including the marginal distribution parameters and the dependence model parameters.
}
\description{
This function fits a Heffernan-Tawn (\code{ht}) model to the given wave data. The marginal distributions are
a mixture model with an empirical distribution below a specified threshold, and the generalised Pareto distribution
(GPD) above the threshold. The dependence structure is the conditional extreme value model proposed in
Heffernan and Tawn (2004).
}
\examples{
# Load data
data(noaa_ww3)

# Fit Heffernan-Tawn model with 0.95 marginal and dependence thresholds
noaa_ht = fit_ht(data = noaa_ww3, npy = nrow(noaa_ww3)/10, p_margin_thresh = 0.95, p_dep_thresh = 0.95)

}
\references{
Heffernan, Janet & A. Tawn, Jonathan. (2004). A Conditional Approach for Multivariate Extreme Values.
Journal of the Royal Statistical Society Series B. 66. 497-546. 10.1111/j.1467-9868.2004.02050.x.
}
