% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_wln.r
\name{fit_wln}
\alias{fit_wln}
\title{Fitting Weibull-log-normal model to wave data}
\usage{
fit_wln(data, npy, weibull_method = "lmom")
}
\arguments{
\item{data}{the wave data in the form a \code{data.table} with wave height \code{hs} and wave period
\code{tp} as columns.}

\item{npy}{the number of data points per year, usually estimated by the number of rows in the
supplied wave data divided by the total period of data coverage (in years).}

\item{weibull_method}{the method for fitting the Weibull distribution to the \code{hs} column in the wave data. Choose
between \code{"lmom"} for L-moment or \code{"mle"} for maximum likelihood estimator. The default option
is \code{"lmom"}.}
}
\value{
A joint distribution object of class \code{wln} containing the key information of a fitted
Weibull-log-normal model, including the three parameters of the Weibull distribution for \code{hs} (as a named
numeric vector) and the six parameters of the conditional log-normal distribution for \code{tp} given \code{hs}
(as an unamed numeric vector, in the order of \eqn{{a_0, a_1, a_2, b_0, b_1, b_2}}).

It is possible to replace one or multiple \code{hs} or \code{tp} parameters in an existing \code{wln} object
(see the examples provided below). The current version of the package does not automatically check the
validity of these user-input parameters. Therefore users are advised to perform the check independently.
}
\description{
This function fits a Weibull-log-normal (\code{ht}) model to the given wave data, such that the wave height
\code{hs} follows a translated (or 3-parameter) Weibull distribution, and the wave period given the wave
follows a conditional log-normal distributuion with the location and scale parameters as functions
of the corresponding \code{hs} value.
}
\details{
The input \code{data} must be a \code{data.table} object with \code{hs} and \code{tp} columns. This can be
generated by reading a CSV file using function \code{\link[data.table]{fread}}.

The formulation of the conditional distribution is given by
\deqn{log(tp | hs=h) ~ N(\mu(h), \sigma(h)^2)}
where the mean and the standard deviation are
\deqn{\mu(h) = a_0 + a_1 h^a_2 and \sigma(h) = b_0 + b_1 exp(h * b_2)}

A two-step regression-based estimation method is used for fitting the conditional distribution of
\code{tp} given \code{hs}. The conditonal means and standard deviations are
estimated over fixed-width bands of \code{hs} in the first step, and regressed onto the above
functions of \code{hs} in the second step.
}
\examples{
# Load data
data(ww3_pk)

# Fit Weibull-log-normal distribution 
wln1 = fit_wln(data = ww3_pk, npy = nrow(ww3_pk)/10, weibull_method = "mle")

# Fit Weibull-log-normal distribution using the alternative method
wln2 = fit_wln(data = ww3_pk, npy = nrow(ww3_pk)/10, weibull_method = "lmom")

# Update the hs parameters for object wln1
wln3 = copy(wln1)
wln3$hs$par[["loc"]] = 0.66
wln3$hs$par[["scale"]] = 2.2
wln3$hs$par[["shape"]] = 1.8
  
}
\references{
Haver, Sverre & Winterstein, Steven. (2009). Environmental Contour Lines: A Method for Estimating Long
Term Extremes by a Short Term Analysis. Transactions - Society of Naval Architects and Marine Engineers. 116.
}
\seealso{
\code{\link{fit_ht}}, \code{\link{sample_jdistr}}
}
