% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_jdistr.r
\name{sample_jdistr}
\alias{sample_jdistr}
\title{Drawing a sample from a fitted joint distribution}
\usage{
sample_jdistr(jdistr, sim_year)
}
\arguments{
\item{jdistr}{a \code{wln} or \code{ht} joint distribution object}

\item{sim_year}{the period of the simulate data in equivalent number of years}
}
\value{
a \code{data.table} object with two columns \code{hs} and \code{tp}
}
\description{
Creates a plot of the crayon colors in \code{\link{brocolors}}
}
\examples{

}
